dashboard added

+++ src/components/Booking/BookingDetails.js
@@ -1,30 +1,46 @@
 import React from 'react';
 import { Button } from 'react-bootstrap';
+import { Link } from 'react-router-dom';
 import useAuth from '../Hooks/useAuth';
 import useBooking from '../Hooks/useBooking';
 
 const BookingDetails = (props) => {
-    const {booking, setBooking} = useBooking();
-    const { user} = useAuth();
+    const { booking, setBooking } = useBooking();
+    const { user } = useAuth();
     const { name, email, phone, address, date, packageName, price, duration, _id } = props.booked
-    
+
+
+
     const handleDeleteBooking = id => {
-        const url = `https://radiant-earth-20543.herokuapp.com/bookings/${id}`
-        fetch(url, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0)
-                    alert('Deleted Successfully')
-                const remainingBookings = booking?.filter(booked => booking._id !== id)
-                window.location.reload();
-                setBooking(remainingBookings);
-            });
+        const proceed = window.confirm('Are you sure, you want to delete?')
+        if (proceed) {
+            const url = `https://radiant-earth-20543.herokuapp.com/bookings/${id}`
+            fetch(url, {
+                method: 'DELETE',
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0)
+                        alert('Deleted Successfully')
+                    const remainingBookings = booking?.filter(book => book._id !== id)
+                    window.location.reload();
+                    setBooking(remainingBookings);
+                });
+        }
     }
 
     return (
         <div>
+            {
+                user.email && <Link to="/mybookings" ><Button className="ms-3 text-white" variant="success">My Bookings</Button></Link>
+            }
+
+            {
+                user.email && <Link to="/manageallbookings" ><Button className="ms-3 text-white" variant="success">Manage All Bookings</Button></Link>
+            }
+            {
+                user.email && <Link to="/mybookings" ><Button className="ms-3 text-white" variant="success">Add A New Service</Button></Link>
+            }
             {
                 user.email === email && <div>
                     <p>{name}</p>
